+++ fidget/src/render/mod.rs
@@ -5,6 +5,7 @@
 use crate::{
     eval::{BulkEvaluator, Function, Trace, TracingEvaluator},
     shape::{Shape, ShapeTape},
+    Error,
 };
 use std::sync::Arc;
 
@@ -194,5 +195,93 @@ impl<F: Function> RenderHandle<F> {
 
         // Do this step last because the evaluators may borrow the shape
         shape_storage.extend(self.shape.recycle());
+    }
+}
+
+/// Container representing an ordered, checked list of tile sizes
+///
+/// This object wraps a `Vec<usize>`, guaranteeing three invariants:
+///
+/// - There must be at least one tile size
+/// - Tiles must be ordered from largest to smallest
+/// - Each tile size must be exactly divisible by subsequent tile sizes
+#[derive(Debug, Eq, PartialEq)]
+pub struct TileSizes(Vec<usize>);
+
+impl std::ops::Index<usize> for TileSizes {
+    type Output = usize;
+
+    fn index(&self, i: usize) -> &Self::Output {
+        &self.0[i]
+    }
+}
+
+impl TileSizes {
+    /// Builds a new tile size list, checking invariants
+    pub fn new(sizes: &[usize]) -> Result<Self, Error> {
+        if sizes.is_empty() {
+            return Err(Error::EmptyTileSizes);
+        }
+        for i in 1..sizes.len() {
+            if sizes[i - 1] <= sizes[i] {
+                return Err(Error::BadTileOrder(sizes[i - 1], sizes[i]));
+            } else if sizes[i - 1] % sizes[i] != 0 {
+                return Err(Error::BadTileSize(sizes[i - 1], sizes[i]));
+            }
+        }
+        Ok(Self(sizes.to_vec()))
+    }
+    /// Returns the length of the tile list
+    #[allow(clippy::len_without_is_empty)]
+    pub fn len(&self) -> usize {
+        self.0.len()
+    }
+
+    /// Returns an iterator over tile sizes (largest to smallest)
+    pub fn iter(&self) -> impl Iterator<Item = &usize> {
+        self.0.iter()
+    }
+
+    /// Returns the last (smallest) tile size
+    pub fn last(&self) -> usize {
+        *self.0.last().unwrap()
+    }
+
+    /// Gets a tile size by index
+    pub fn get(&self, i: usize) -> Option<usize> {
+        self.0.get(i).copied()
+    }
+
+    /// Returns the data offset of a global pixel position within a root tile
+    ///
+    /// The root tile is implicit: it's set by the largest tile size and aligned
+    /// to multiples of that size.
+    #[inline]
+    pub(crate) fn pixel_offset(&self, pos: nalgebra::Point2<usize>) -> usize {
+        // Find the relative position within the root tile
+        let x = pos.x % self.0[0];
+        let y = pos.y % self.0[0];
+
+        // Apply the relative offset and find the data index
+        x + y * self.0[0]
+    }
+}
+
+/// Hints for how to render this particular type
+pub trait RenderHints {
+    /// Recommended tile sizes for 3D rendering
+    fn tile_sizes_3d() -> TileSizes;
+
+    /// Recommended tile sizes for 2D rendering
+    fn tile_sizes_2d() -> TileSizes;
+
+    /// Indicates whether we run tape simplification at the given cell depth
+    /// during meshing.
+    ///
+    /// By default, this is always true; for evaluators where simplification is
+    /// more expensive than evaluation (i.e. the JIT), it may only be true at
+    /// certain depths.
+    fn simplify_tree_during_meshing(_d: usize) -> bool {
+        true
     }
 }