+++ CHANGELOG.md
@@ -18,6 +18,7 @@
     - The previous `fidget::shape::Bounds` type is removed
     - Remove `fidget::render::render2d/3d` from the public API, as they're
       equivalent to the functions on `ImageRenderConfig` / `VoxelRenderConfig`
+- Move `RenderHints` into `fidget::render`
 
 # 0.3.3
 - `Function` and evaluator types now produce multiple outputs